requirements.txt
click>=8.0
python-dotenv
Pillow>=10.0
pillow-heif
flask>=3.0
werkzeug
Jinja2>=3.0
google-auth
google-auth-oauthlib
requests
tqdm
pypdf
pytest
